Finalizing strict mode fixes
