+++ front/src/Phaser/Reconnecting/ReconnectingScene.ts
@@ -4,7 +4,7 @@ import {TextInput} from "../Components/TextInput";
 import {ClickButton} from "../Components/ClickButton";
 import Image = Phaser.GameObjects.Image;
 import Rectangle = Phaser.GameObjects.Rectangle;
-import {PLAYER_RESOURCES} from "../Entity/PlayableCharacter";
+import {PLAYER_RESOURCES} from "../Entity/Character";
 import {cypressAsserter} from "../../Cypress/CypressAsserter";
 import Sprite = Phaser.GameObjects.Sprite;
